+++ jkube-kit/generator/java-exec/src/main/java/org/eclipse/jkube/generator/javaexec/JavaExecGenerator.java
@@ -155,7 +155,6 @@ public class JavaExecGenerator extends BaseGenerator {
       addJolokiaPort(buildBuilder);
       addPrometheusPort(buildBuilder);
 
-      addLatestTagIfSnapshot(buildBuilder);
       addTagsFromConfig(buildBuilder);
       buildBuilder.workdir(getBuildWorkdir());
       buildBuilder.entryPoint(getBuildEntryPoint());